improve test cases

+++ examples/integration-test.py
@@ -6,18 +6,19 @@ import shutil
 from common.config import Config as C
 from main import main
 
-need_gen = False
-for d in C.TRAIN_DAYS:
-    if not os.path.exists(f"{C.DATA_DIR}/{d}"):
-        need_gen = True
-        break
+if "OMPI_COMM_WORLD_RANK" not in os.environ or os.environ["OMPI_COMM_WORLD_RANK"] == "0":
+    need_gen = False
+    for d in C.TRAIN_DAYS:
+        if not os.path.exists(f"{C.DATA_DIR}/{d}"):
+            need_gen = True
+            break
 
-if need_gen:
-    import gen_example_data
+    if need_gen:
+        import gen_example_data
 
-    gen_example_data.main()
+        gen_example_data.main()
 
-# remove existing model
-shutil.rmtree(C.MODEL_DIR, ignore_errors=True)
+    # remove existing model
+    shutil.rmtree(C.MODEL_DIR, ignore_errors=True)
 
 main()
